Update en_us.json (#52104)
* Update en_us.json

* Update en_us.json

I noticed the original commit rearranged a bunch of strings. I chatted with Katie and got her permission to revise it so the strings were mostly back in alphabetical order. Hopefully this makes the diff look smaller - ultimately: this is just 2 revised and 1 brand new string, but the diff made it look way bigger.

* Update en_us.json

Alphabetize keys

---------

Co-authored-by: Dan <68714964+dancodedotorg@users.noreply.github.com>
Co-authored-by: bencodeorg <ben@code.org>