Increase Health Check Timeout to 5 Minutes
We currently wait for a maximum of 60 seconds for frontend servers to
come online, but recent changes in
https://github.com/code-dot-org/code-dot-org/pull/52313 mean that we now
get to this point in the build process while the web servers are still
restarting (rather than blocking earlier in the build process until we
detect that they're up and running).

Both because of that and because curl's exponentional backoff logic mean
that we only actually get in 5 or 6 retries in 60 seconds rather than
the full 10 we have configured, increase timeout to give the web servers
a better chance of being fully started up.