Set `Rails/Date` `EnforcedStyle` to "strict"
And fix the resulting violations. Follow-up to https://github.com/code-dot-org/code-dot-org/pull/50760, which enabled the rule with its default configuration and updated all instances of `Date.today` to `Time.zone.today`; this PR also updates instances of `Date.(current|yesterday|tomorrow)`.

I think it's entirely optional for us to take on this additional level of linting beyond what Rubocop provides by default, but I quite like the consistency of this approach. I'm open to suggestions on whether or not we want to actually enable the rule in the long run, but I do think we want to apply these fixes regardless.