Reinstall Ruby when Target Install Directory Changes
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/50661; our current logic will only try to embark on the time-consuming process of building and installing a new version of ruby when the version we're targeting changes. Unfortunately, that means we don't do anything when we try to change the location that ruby installs to unless we also change the version.

In order to support our ability to update the Ruby installation directory, update our gating logic to also examine that.