Fixes a flaky CPA test.
Adds a UI test step that can detect when a page was reloaded. This is
needed because the test in question loads the same page at the same URL.
Checking for the URL or an element on the page is not good enough since
it would be ambiguous between the original page and the reloaded page.
There is no other way to wait until the text being check could possibly
be updated.