Fix Violations of and Reenablel `Lint/OrAssignmentToConstant`
> Checks for unintended or-assignment to a constant.
>
> Constants should always be assigned in the same location. And its value should always be the same. If constants are assigned in multiple locations, the result may vary depending on the order of `require`.

Fix applied automatically with `bundle exec rubocop --auto-correct-all --only Lint/OrAssignmentToConstant`

This may well be a Chesterton's Fence situation, because in neither case is it clear to me why we are using conditional assignment for these constants in the first place. I can't think of any reason why there would be, for the reasons pointed out in the description of the cop itself; I think it's likely these are just oversights, but want to call out that I'm not certain about that.

- https://docs.rubocop.org/rubocop/cops_lint.html#lintorassignmenttoconstant
- https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Lint/OrAssignmentToConstant