Update `aws-sdk-core` Gem to 3.180
Update applied automatically with `bundle update aws-sdk-core`. Note that we could have used the `--conservative` flag to do so without updating the other two libraries in this changeset, but as this is a small, safe update anyway I see no reason not to include them all.

The motivation for this update is mostly speculative, to see if reinstalling the gem helps with the strange issues we're seeing on test whenthis gem tries to invoke the `time` gem.

See [slack thread](https://codedotorg.slack.com/archives/C0T0PNTM3/p1690427314013019) for more context.