Explicitly Install `ruby-build` Dependencies
`ruby-build` requires `libz` to be installed in order for the "install ruby with ruby build" step to succeed. We install that library in [another cookbook](https://github.com/code-dot-org/code-dot-org/blob/34175761b482c6a4df605c1823389a04e431d79c/cookbooks/cdo-tippecanoe/recipes/default.rb#L1), which presumably is why this has not been an issue on any of our persistent managed servers, but don't install it before building ruby on a new server.

It's not clear why this has not been a problem on adhocs.