Install `pdftk-java` on Test Server
Like we do on staging; otherwise, our pegsus tests fail with "Expected 'pdftk' to be installed.", and can't effectively test our PDF generation logic.

Note that the test server previously did have this dependency installed, despite the absense of automatic installation logic. It's not clear exactly how this happened, but we are assuming it was a manual effort.