Install `pdftk-java` on Test Server (#53213)
* Install `pdftk-java` on Test Server

Like we do on staging; otherwise, our pegsus tests fail with "Expected 'pdftk' to be installed.", and can't effectively test our PDF generation logic.

Note that the test server previously did have this dependency installed, despite the absense of automatic installation logic. It's not clear exactly how this happened, but we are assuming it was a manual effort.

* target adhocs for pdftk installtion, both to test this functionality in the short term and to regularly verify it in the long term since staging and test don't often get recreated

* target pdftk 3.0.9 since 3.1.1 apparently doesn't exist

* updated checksum; verified both with sha256sum and at https://packages.ubuntu.com/focal/all/pdftk-java/download