Explicitly Disable `Performance/Detect`
> This cop is used to identify usages of `select.first`, `select.last`, `find_all.first`, and `find_all.last` and change them to use `detect` instead.
>
> `ActiveRecord` compatibility: `ActiveRecord` does not implement a `detect` method and `find` has its own meaning. Correcting ActiveRecord methods with this cop should be considered unsafe.

Because this cop is unsafe to apply to ActiveRecord methods and because those methods represent the majority of our existing violations of this cop, I think it's safest for us to disable this cop across the whole project. We could alternatively enable it just for directories other than `dashboard`, which for a `Performance/` cop might be justified, but I chose to go with the simpler option.

What do y'all think?

- https://www.rubydoc.info/gems/rubocop/0.43.0/RuboCop/Cop/Performance/Detect
- https://docs.rubocop.org/rubocop-performance/cops_performance.html#performancedetect