Update `level_file_path` to Consistently Return a Pathname (#53345)
* Update `level_file_path` to Consistently Return a Pathname

Right now, `level_file_path` will return the path as a Pathname when it does not already exist, and as a String if it does. This is because we invoke `default_level_file_path` when it doesn't, which returns a Pathname, and when it does we return the result of `Dir.glob(...).first`, which is a String.

To make this method both more consistent internally and with the results of `default_level_file_path`, update the 'file already exists' case to also return a Pathname.

* update test to reflect new more-consistent functionality