Update some Hardcoded Level File Paths
As part of our ongoing work to reorganize the level files into smaller subdirectories. The specific existing level files referred to by these tests got reorganized on levelbuilder last night, and the DTT started failing this morning because the files are no longer located at those paths.

In the long term, we should probably update this test to not be using actual site data like it is, but in the short term simply updating the hardcoded paths to point to the new location should work to get everything passing again.

I also grepped our test directory for other instances of `config/scripts`, and confirmed that all instances are testing content that still belongs in that directory.