Let DOTDs DTP Even if PR Already Exists (#53735)
* Let DOTDs DTP Even if PR Already Exists

Right now in the dotd script, if you initiate a DTP and then say "no" when it asks you if you want to merge it, it will break when you attempt to re-initiate the DTP because it doesn't expect the git branch or github PR to have already been created. This change updates the script to recognize and reuse previously-created entities.

* add warning when reusing branch

* Verify commit sha of dtp candidate branch before reusing