Allows account email to send parent permission if parent created account.
This adds the ability to set the parent_email via the registration
process in the UI tests. Adds steps to 'press keys' to input email since
the component reacts to the input events (which are bypassed by just
setting the value directly.)

This swaps out the 'student-email' data for the LockoutPanel component
in favor of the more specific 'disallow-email' tag. This tag is, then,
empty when the parent created the account so that the account email is
allowed when submitting the form via the front-end component.