Set `EnforcedStyle: refute` for `Rails/RefuteMethods` (#53774)
* Set `EnforcedStyle: refute` for `Rails/RefuteMethods`

Fix applied automatically with `bundle exec rubocop --autocorrect-all --only Rails/RefuteMethods`

With this configuration option, this rule enforces:

> Use `refute` methods instead of `assert_not` methods.

Note that as an alternative, we could enable this rule with the default configuration, which enforces the use of `assert_not` methods over `refute` methods. This option results in a *much* larger change to our codebase; 1,998 offenses in total, versus 217 offenses with this configuration option. It's also not clear to me why `assert_not` is preferred over `refute`; personally, I think `refute` results in more-readable test logic, with a lot fewer double- and triple-negatives.

- https://docs.rubocop.org/rubocop-rails/cops_rails.html#railsrefutemethods
- https://www.rubydoc.info/gems/rubocop-rails/RuboCop/Cop/Rails/RefuteMethods

* autocorrect some newly-added violations