Modify Notification to allow for a new co-teacher invite notification. (#54068)
In order to add the new co-teacher workflow, we want to add a co-teacher invite notification. This was designed to have a different icon and slightly different styles to existing notification types. This PR adds the following to the Notification object in order to follow the Figma designs:

An optional tooltip that shows text on hover
The ability to add styles and color to the optional buttons
A new 'collaborate' notification type that uses the "users" icon and secondary brand colors

---------

Co-authored-by: Liam CDO <liamcdo@Liams-MBP.localdomain>