Extend Startup Timeout for new Web Services
On the first build following the Ubuntu 20 upgrade, our newly-created web servers are taking a very long time to start up dashboard and pegasus the first time. Speculatively double the timeouts for the individual services from 5 minutes to 10 minutes, and also double the timeout for the frontend as a whole from 10 minutes to 20 minutes.