Return Early from DCDO Get if No Key Set
In our previous implementation, we'd never attempt to invoke `.get` directly on the datastore for keys that hadn't ever been set; we just used `.all` to collect everything that was set and put in all in the cache, then tried to get values off the cache.

With the new implementation, we will actually check the datastore for values. This means that any key that hasn't had an explicit value set will attempt to load a `nil` value from the datastore, which right now for the JSON-backed and in-memory datastores will also log an error in Honeybadger.

Instead, we simply check for the existence of the key in the datastore before trying to parse it as JSON. Note that this is already how it works for the DynamoDB-backed datastore used in production; see [this comparable line](https://github.com/code-dot-org/code-dot-org/blob/02f7d6ea7a50454b7304ff81024ff211b81bd13d/lib/dynamic_config/adapters/dynamodb_adapter.rb#L22) in that version of the adapter.