Automatically Create `.bundle/config` via Chef (#54343)
We at some point stopped creating a user-global bundle config file in the home directory of the managed servers we provision with chef, because it interfered with the creation of cookbooks. But that also means that we are not able to manually run ruby scripts on newly-provisioned servers, including scripts like `bin/start-build` that might be necessary to also run the provisioning scripts.

Instead, we start creating a bundle config file at the root of the project; this should avoid applying it globally in a way that breaks cookbooks, while still enabling regular execution.

TODO: verify that installing the bundle config to the project root rather than the user home is sufficient to avoid re-breaking cookbooks.