Remove Custom `review` Sinatra Verb
Originally added in https://github.com/code-dot-org/pegasus/commit/931341aca779b6bbaf650eaef73d423903a61433, we apparently have the ability to define custom verbs beyond `GET`, `POST`, `PATCH`, etc, and do so in exactly one place in our codebase. Because the verb we define is not actually a verb used by HTTP clients, we have a classic `POST` route that is actually being used on the frontend and which internally simulates a call to the custom route.

This PR proposes a simplification: simply use the `POST` route directly, and remove all the custom verbs logic.

I do want to call out that this PR is a Chesterton's Fence violation. I do not know why this particular fence was put up in the first place, and am therefore not fully confident that it is safe to remove. All I can say is that I've gone looking for a reason why it *should* exist, and have been unable to find one.

TODO: testing. Still trying to figure out how to locally access the view that's generating these POSTs