Update tasks.yml
Adding 3 tables to be included in DMS to support analysis of new AI features.

I've added to the pii section because there is free response (teacher feedback) data that may be included in one of the tables and I want to keep the other related tables co-located in the same schema.