Fix Live Reloading of new DynamicConfig Cache in Test
In the test environment, it's not only possible but expected for the underlying shared cache to be cleared out while the environment is still running. In that situation, when we attempt to update Dynamic Config's new local cache from the shared cache before the shared cache has been updated from the long-term data store, we might accidentally set it to the `nil` value and break tests.

We don't expect to ever encounter this in any production environment, and the simple fix is to make sure that when that specific situation occurs we default to something reasonable rather than something which breaks our assumptions.