Use In-Memory Shared Cache for Parallel Tests
Another follow-up to https://github.com/code-dot-org/code-dot-org/pull/54601

When the FeatureModeManager tests are executed with `rake parallel:test` (as they are in the DTT), the flags can now overwrite one another since they are all using the same filesystem-based shared cache. To prevent this, stub out that shared cache for an in-memory one.