Enable `jsx-a11y/aria-role` Only for DOM Nodes
By default, this rule inspects developer-added components as well as generics. Unfortunately, that means that if we define a `role` property representing something other than an ARIA role, the rule will warn on that as well.

Currently, this rule only warns on two instances of a single component, specifically our use of the `CohortCalculator` component in tests:
https://github.com/code-dot-org/code-dot-org/blob/e6095851286f681f3c75e0db64941334ce5c09d9/apps/test/unit/code-studio/pd/application_dashboard/cohort_calculatorTest.js#L17-L21
https://github.com/code-dot-org/code-dot-org/blob/e6095851286f681f3c75e0db64941334ce5c09d9/apps/test/unit/code-studio/pd/application_dashboard/cohort_calculatorTest.js#L52-L56

Our current violations are minor enough that we could easily fix the violations either by renaming the property or adding two minor ignore directives, and then enable this rule in full. I am, however, worried about the potential for future confusion; keeping this rule enabled for developer-created components means we will expect to receive false positives any time we create a component with a property named `role`.

What do y'all think? Is it worth it to keep the rule enabled in full, or does scoping it to just generic components that are using `role` to represent an ARIA role make sense?

https://github.com/jsx-eslint/eslint-plugin-jsx-a11y/blob/main/docs/rules/aria-role.md