Use Ubuntu 20 in All Managed Environments
Alternatively, "Update Levelbuilder Environment to use Ubuntu 20".

This PR removes the per-environment logic we added to facilitate a gradual rollout of Ubuntu 20. Currently, the only managed environment not yet using Ubuntu 20 is Levelbuilder, which this PR will update.

After this, the only long-running server of ours yet to be updated to Ubuntu 20 is the gateway server, which we plan to update in-place rather than via CloudFormation replacement.

Previous work:

- https://github.com/code-dot-org/code-dot-org/pull/54269
- https://github.com/code-dot-org/code-dot-org/pull/54271
- https://github.com/code-dot-org/code-dot-org/pull/54272