Rename `levelbuilder-staging` to `levelbuilder`
Specifically, update all references to `levelbuilder-staging` in the codebase to instead reference `levelbuilder`. This is in response to the fact that the new version of the server created as part of the Ubuntu 20 upgrade was created with the name `levelbuilder` instead of the name `levelbuilder-staging`

We believe the server previously used the `-staging` suffix because it was manually given that name when it was originally created, and the name was preserved when the server was imported into CloudFormation. Because this upgrade was the first time that the CloudFormation stack attempted to create the server from scratch, it's only now that we see the suffixless version of the name.

We could alternatively update the underlying template to attempt to preserve the exisiting name, but the suffix has been a common stumbling block and source of confusion, so we've decided to instead embrace this opportunity for standardization.