Undo Custom DNS Patch for Ubuntu 18
[The underlying bug](https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1805183) that prompted us to add this patch in the first place was fixed in Ubuntu 19; now that we're on Ubuntu 20, we no longer need the patch. However, we can't simply remove the logic that adds the patch, or our existing persistent managed servers will retain the patch as an artifact which may someday introduce unexpected differences between old and new servers. Instead, this PR will update the contents of the patched file on all existing servers to match what the contents would be on a newly-provisioned server without the patch applied.

This version of the file was obtained from an adhoc that did not have the patch applied in the first place, and I confirmed that the adhoc works just fine. I am unfortunately not sure how to attempt to reproduce the original bug this was intended to fix, so I can't be 100% certain that the bug no longer manifests on this version of Ubuntu, and am choosing to trust the [original bug report](https://bugs.launchpad.net/ubuntu/+source/systemd/+bug/1805183) when it says this issue has been fixed in the version of Ubuntu we are now using.

Once all existing servers have been updated with this "unpatched patch", we can then [remove the patching logic entirely](https://github.com/code-dot-org/code-dot-org/pull/54427).