Update setup instructions for WSL (#55573)
Cleanup and clarification of SETUP.md instructions after going through a WSL setup on a new machine. Clarified the path needed for Chrome bin to be wrapped in quotes and have escape characters for the spaces in the path. Without both, the bin/aws_access script fails due to browser failing to open.