Adding FAIL to the common build error text
The description as it was, was not very obvious as I went into the logs while on DOTD.  Not knowing which was the error slowed me knowing what to search for.  Once I knew which error I could search for "Generated different package for same contents" in the DOTD document.

This slight change, I feel will make this much more readable in this error case.