Actually Use `head` Field in DOTD Script to Filter PRs
Right now, we attempt to identify the DTP PR by filtering by both `base` and `head`. Unfortunately, the value we are currently passing for `head` is just the name of the branch, and this endpoint wants us to also include an organization identifier or it will ignore that value and filter only by `base`.

This hasn't been a problem before because it's usually the case that the only PR open against `production` is the DTP. But if, theoretically, we were to [have an open hotfix PR waiting to be reviewed when the DOTD runs DTP](https://codedotorg.slack.com/archives/C0T0PNTM3/p1707518933025489), then that hotfix PR would be incorrectly identified as the DTP PR.

Tested a few `Octokit.pull_requests` calls manually:

```ruby
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging', head: 'mysql-8-in-chef'}).count
=> 30
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging'}).count
=> 30
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging', head: 'dtp_candidate_92f7b0c9'}).count
=> 30
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging', head: 'hamms:mysql-8-in-chef'}).count
=> 0
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging', head: 'code-dot-org:mysql-8-in-chef'}).count
=> 1
[development] dashboard > Octokit.pull_requests(GitHub::REPO, {base: 'staging', head: 'code-dot-org:dtp_candidate_92f7b0c9'}).count
=> 0
```

- https://docs.github.com/en/rest/pulls/pulls?apiVersion=2022-11-28#list-pull-requests--parameters