Adds more detailed student/teacher info admin views.
This adds a Assume Identity button to the per-student and per-teacher
view.

This adds a listing of sections that the student (or teacher) is
involved in including a link to the teacher info and their email.

This adds a listing of authentications on this view for quick viewing of
the authentication types the account has available.