Fix warning on `yarn build`: WARNING in ./node_modules/asn1.js/lib/asn1/api.js 21:12-42 Module not found: Error: Can't resolve 'vm' in '/Users/seth/src/code-dot-org/apps/node_modules/asn1.js/lib/asn1'
BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
        - add a fallback 'resolve.fallback: { "vm": require.resolve("vm-browserify") }'
        - install 'vm-browserify'
If you don't want to include a polyfill, you can use an empty module like this:
        resolve.fallback: { "vm": false }