[LABS-282] Improve UI/instructions how to update the Firmata firmware for CP Express using alternative path (#56340)
Switching the CP Express to bootloader mode and re-installing the firmata firmware is not an alternative pathway like we previously assumed. It must be done every time one switches back to Code.org after using Microsoft MakerCode. With that in mind, we are including the instructions in our main maker/setup instructions.

Shoutout to @fisher-alice for providing image assets and being a hand model!