feat(lms): do not sync if not section owner
Previously, any teacher on the LMS for a course will trigger course
syncing even if the section owner may not actually want syncing to
occur.

This PR updates the behavior to only trigger course syncing if this is
the section owner. The "Sync Course" button is also hidden unless you
are the section owner.