feat(lms): do not sync or create sections if on the LMS, the user is not an instructor
Previously, the sync_course logic will operate regardless of the user's
role on the LMS. This resulted in some cases where synchronization
occurred when not a teacher (for example, an administrator inadvertently
synchronizing a course, or a student creating a course). When a
non-instructor creates a course, the section becomes "owned" by the
non-instructor potentially locking out the actual instructor.