feat(signup): move emailfrom backend session cache to browser session storage
Currently, the signup flow is a two page process where email and
password is collected on the first page and additional metadata on the
second. This results in the email being stored in the backend session cache.

Instead, we can temporarily store the email in the browser session
storage. The handling of invalid or missing states will be consistent with the
existing backend behavior (cancel and return to first page).