Adds feature to EditorAnnotator to color lines and tooltips and uses that for AI TA.
The AI responses will be annotated with new styling and the consistent
blue coloring as found elsewhere instead of the bright yellow 'school
highlighter' coloring. A new 'i' icon to replace the built-in one and
the little AI bot icon to adorn the tooltip itself.

Long tooltips are also formatted to have a max width. This styling is
always applied to annotation tooltips. EditorAnnotator has extra
functionality to allow adding more styling. Unfortunately, since this is
applied to the element specifically, the styling is written in the
LearningGoals component source instead of the CSS files.