Automatically Fix `import/order` Violations for Instructions Templates
Specifically, I reenabled this newly-added rule for the directories `src/templates/instructions/` and `test/unit/templates/instructions`, and ran `yarn lint:fix` to update all files within those directories.

I plan to manually review all file changes in the GitHub UI to identify any oddities, but otherwise am relying on existing tests to verify that this lint fix does not result in any changes to functionality.

Open to feedback on whether or not it makes sense to bundle `src` and `test` directories together like this; it has a certain aesthetic appeal, but I'm not sure there's any practical reason to do so.