Instrument Cookie-based Session Storage
In preparation for moving our session storage out of the user's browser and into some centralized data store, we'd like to get an idea of just how many active sessions we actually have, and of how much data is being stored in them.

To that end, this PR adds some simple instrumentation to the read, write, and delete operations on the CookieStore which will publish metrics to AWS CloudWatch.