Split Up Name Argument in `Cdo::Metrics.put` (#57376)
* Split Up Name Argument in `Cdo::Metrics.put`

Our `Cdo::Metrics` helper currently has three methods for logging metrics data; `push` and `put_metric` both expect you to define the namespace and metric name as separate strings, but `put` expects you to put them both in a single string separated by a `/`. Unfortunately (and I think unintentionally), this precludes us from being able to use namspaces which themselves have a `/` in them. Not only do `push` and `put_metric` allow that kind of formatting, but we are indeed [taking advantage of that functionality](https://github.com/code-dot-org/code-dot-org/blob/b3dcd016d5cb17c696e0fc9e6436f87ad5a1e2a2/bin/cron/report_activejob_metrics#L7-L12)

Because I don't think we have a specific reason to disallow `/` in the namespace and because the specific implementation of the `put` method does not appear to be actually saving us from a significant amount of complexity, I propose that we refactor the method to be more in line with the other methods in the class.

Thoughts? Concerns? Any specific context as to why we originally implemented it this way that I'm missing?

* update one more test to reflect new method signature

* update an invocation of Cdo::Metrics.put that I missed