Instrument Cookie-based Session Storage (#57357)
* Instrument Cookie-based Session Storage

In preparation for moving our session storage out of the user's browser and into some centralized data store, we'd like to get an idea of just how many active sessions we actually have, and of how much data is being stored in them.

To that end, this PR adds some simple instrumentation to the read, write, and delete operations on the CookieStore which will publish metrics to AWS CloudWatch.

* misc tweaks

* Split Up Name Argument in `Cdo::Metrics.put`

Our `Cdo::Metrics` helper currently has three methods for logging metrics data; `push` and `put_metric` both expect you to define the namespace and metric name as separate strings, but `put` expects you to put them both in a single string separated by a `/`. Unfortunately (and I think unintentionally), this precludes us from being able to use namspaces which themselves have a `/` in them. Not only do `push` and `put_metric` allow that kind of formatting, but we are indeed [taking advantage of that functionality](https://github.com/code-dot-org/code-dot-org/blob/b3dcd016d5cb17c696e0fc9e6436f87ad5a1e2a2/bin/cron/report_activejob_metrics#L7-L12)

Because I don't think we have a specific reason to disallow `/` in the namespace and because the specific implementation of the `put` method does not appear to be actually saving us from a significant amount of complexity, I propose that we refactor the method to be more in line with the other methods in the class.

Thoughts? Concerns? Any specific context as to why we originally implemented it this way that I'm missing?

* refactor cookie store instrumentation to better reflect both anticipated usage and desired data

* update one more test to reflect new method signature

* wrap `log_session_info` in an error handler

So that even if this new logic messes up, we won't break session cookies

Co-authored-by: Seth Nickell <seth.nickell@code.org>

* log exception to honeybadger if session info logging fails

* CamelCase rather than snake_case metrics names

* update TODO with a link to relevant PR

---------

Co-authored-by: Seth Nickell <seth.nickell@code.org>