Fixes issues with malformed evidence strings.
Since evidence is its own column, we can get away with not having some
of the information and falling back to what is there. If the code isn't
there, we still annotate with the message for those lines.

If the message isn't there, then we annotate with the observations
column instead. Here, we only use the line numbers to mark the
highlights. The evidence listing is just a bulleted list of sentences
provided in the observations column.

We also allow for partial matches of code. The AI sometimes truncates
the code down (e.g. 'if (something) { ... }') and in those cases we want
to gracefully fall back to highlighting the parts we match. It's not
ideal.