Write Session Data to Redis
Notably, however, do not yet begin relying on Redis for *reading* session data.

We want to get an accurate evaluation of what real-world usage of this
new Redis cluster will look like, and also benefit from some load
testing in advance of an actual switchover. To that end, this PR
proposes that we begin writing all session data to Redis as a side
effect of using the existing cookie-based store. This new functionality
will be gated behind a DCDO flag and wrapped in error handling logic to
minimize the chances of this experiment affecting the user experience.

Follow-up to https://github.com/code-dot-org/code-dot-org/pull/57558