chore(jest): port appLabTest to jest
This test results in certain calls to initApp which calls out to API
endpoints. In Karma, these APIs would reject and fail silently but jest
will error on them. Stub out this API call.

Additionally, the dropletConfig that was being passed needed to be an
immutable, therefore clone the object.