fix(CoteacherTable): remove orphaned style
This span references a class name does not appear to exist in the
included scss file. This resulst in an undefined class to be set to the
span resulting in an error in more strict parsing libraries. Since the
tag does not appear to reference anything, removing the style to avoid
injecting an undefined into the DOM.