Write Session Data to Redis (#57562)
* Provision new Redis-based Session Store

The first step in our efforts to transfer user session data from cookies
to a centralized data store is to provision the data store in
Cloudformation, so we can begin writing data to it.

We have quite a few options for how exactly we want to provision this,
but broadly speaking we can choose either Serverless or On-Demand nodes,
and if we choose On-Demand we also must choose both the size and
quantity of nodes we want.

According to Google Analytics, we've had 28M users over the last 90
days, and 84M over the last year, which works out to between 45 and 60
million users on average over an arbitrary 200-day period (200 days
being our current session duration configuration). At ~300 bytes per
session on average, that's 13.5-18 GB of data we'd ideally like to
store.

Storing 18GB of data in Severless ElastiCache would cost ~$1620/month;
even 13GB would still be ~$1170/month.  Alternatively, we could use
three of the smallest memory-optimized cache node (cache.r7g.large) and
get 39.21GB of storage for ~$480/month, or even just use two for 26.14GB
of storage at ~$320/month. An even more targeted solution could be to
use a larger number of smaller nodes; 6 cache.t4g.medium nodes would
allow us to store 18.5GB for ~$280/month.

Throughput is less of an issue for us than storage. According to
CloudWatch, we average ~800M requests per week, 15M/hour, and under
4,200/second. At an average of 300 bytes stored per session and one GET
plus one PUT per request, that's 2.5 Gigabits of needed network
performance for on-demand nodes (well under the 12.5 promised by
cache.r7g.large nodes), or ~$25/month in ECPUs for a Serverless model.

This PR proposes we provision a single cluster consisting of three
cache.r7g.large nodes, with encryption, automatic failover, and
multi-zone availability all enabled. That should give us plenty of room
for growth without too much unnecessary expense. The next step in our
broader efforts to enable this feature is to begin writing data to redis
so we can get a more accurate idea of actual usage, at which point we
can revisit this decision.

* Write Session Data to Redis

Notably, however, do not yet begin relying on Redis for *reading* session data.

We want to get an accurate evaluation of what real-world usage of this
new Redis cluster will look like, and also benefit from some load
testing in advance of an actual switchover. To that end, this PR
proposes that we begin writing all session data to Redis as a side
effect of using the existing cookie-based store. This new functionality
will be gated behind a DCDO flag and wrapped in error handling logic to
minimize the chances of this experiment affecting the user experience.

Follow-up to https://github.com/code-dot-org/code-dot-org/pull/57558

* remove instrumented cookie store; we will instead get useful data by inspecting the Redis cluster to which we are mirroring data

* also update some references to STDIN

* Revert "also update some references to STDIN"

This reverts commit b3a73702014775cb789069e808d0447a8ae89384.

* use secure redis connection ('rediss://') instead of regular ('redis://')

* temporarily provision session store in non-frontend environments, for testing

* also temporarily configure session store in non-frontend envrionments, for testing

* remove duplicate declaration

* return session store to frontends-only

* add TODO and use versioned rather than commit-based GH link