Mirror Session Deletion to Redis Store
Follow-up to https://github.com/code-dot-org/code-dot-org/pull/57562,
which started writing session data to Redis with the goal of giving us a
chance to analyze real-world usage in advance of a switch to actually
start using Redis as the source of truth.

For the most part, that's working, but there are currently a couple of
issues with that implementation which both affect the total amount of
data being stored, and render that information inaccurate.

First, we aren't deleting old sessions from Redis, so aren't capturing
the extent to which users logging out will reduce total data storage. To
resolve that, this PR simply extends the
[`delete_session`](https://github.com/rails/rails/blob/v6.1.7.7/actionpack/lib/action_dispatch/middleware/session/cookie_store.rb#L63-L68)
method to begin mirroring deletions in Redis the same as we mirror
writes.

Second, we're actually writing more data than we need to; I mistakenly
thought we were going to need to manually serialize and deserialize the
Ruby Hashes that our session data is stored in, and chose to do so with
`Marshal`, which is much less space-efficient than simply passing the
Hash directly to Redis to store. In my local testing, a Hash for a
freshly-logged-in session is 168 bytes in memory, but the `Marshal`
serialization takes 424 bytes.

Simplifying the write logic and enabling deletion logic should improve
the accuracy of our metrics.