fix(lms): update the LtiResourceLinkId if the passed rlid changes
In certain situations, the LtiResourceLink could have changed resulting
in a mismatch when a teacher attempts to sync via the Manage Students
page. When this occurs, the LMS responds with an access error for the
given resource (since the resource could have been deleted).

To fix this, update the LtiCourse resource_link_id field with the latest
value available from the SSO sync process.