Use Lowercase Value for `ClusterMode` (#58292)
During the DTP which attempted to upsize the Redis cluster we use for
session storage (see
https://github.com/code-dot-org/code-dot-org/pull/58262), the
CloudFormation stack update [failed with the error message "Cluster mode
updates are not supported while attempting to update additional
properties"](https://codedotorg.slack.com/archives/C03CK8FGX/p1714167673511629)

According to [this
answer](https://repost.aws/questions/QUj6IF3l-qSsKKkD3yxY_EDw/elasticache-cluster-mode-updates-are-not-supported-while-attempting-to-update-additional-properties#ANrtmd6mpHRYaDBcihifZyhg)
on AWS re:Post, it seems like this is an issue with CFN documentation.
Specifically, the documentation claims that the `ClusterMode` property
should have a value starting with an uppercase letter; either `Enabled`
or `Disabled`, but in reality it expects an entirely lowercase value;
either `enabled` or `disabled`. Because of that discrepancy,
CloudFormation thinks we are attempting to update `ClusterMode` when
we're really just trying to update `CacheNodeType`

To resolve, simply use the expected rather than documented value.