feat(email): add script to scan for legacy partial registration keys
This oneoff script scans the shared memcached instance for keys that do
not have a TTL and matches one of the partial registration keys. In
`DRY_RUN=true`, the keys will be printed out. In `DRY_RUN=false` the
keys will be removed.

This script only works for memcached caches as filesystem caches aren't
utilized in production and thus not stored.