Add DCDO Flag to Disable Session Migration
In preparation for eventually removing this code entirely, add a DCDO
flag that will allow us to dynamically disable the custom functionality
provided by this code, to build our confidence that removal will be
safe.

Note this is probably not strictly necessary; my expectation is that we
should be able to just switch to the generic RedisStore implementation
directly without any user disruption. This flag represents an abundance
of caution, and I don't think there's any reason not to move cautiously.