fix(statsig): add statsig stable id to the rails session
Normally, statsig requires a user id to track a user's journey. However,
anonymous users will inherently not have a user id because the user has
not yet been created. Instead, use the 'stable id' feature which is a
random uuid generated alongside the rails session. The reason why a new
uuid is generated is to prevent leaking the rails session id to external
parties.

A new method `log_statsig_event_with_session` has been created to ensure
most calls are made with the user and only in unique circumstances (such
as tracking an anonymous user) should this method be used.