Use `template` EnforcedStyle for `Style/FormatStringToken`
The default style for this cop is `annotated`, which looks like
`format('%<greeting>s', greeting: 'Hello')`

Currently in our codebase, we consistently use the `template` style,
which looks like `format('%{greeting}', greeting: 'Hello')`; similar to
the style used by classic string interpolation. I'm not sure whether
this was intentional or if we simply defaulted to this style since most
of us are likely more familiar with string interpolation than with the
format method, but given how consistent we are (and how many violations
there are with the default style), I think it makes sense to codify this
behavior.

- https://www.rubydoc.info/github/bbatsov/RuboCop/RuboCop/Cop/Style/FormatStringToken
- https://docs.rubocop.org/rubocop/cops_style.html#styleformatstringtoken